requirements.txt
aiohttp==3.8.6
aiosignal==1.2.0
aiosqlite==0.19.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
exceptiongroup==1.1.3
frozenlist==1.4.0
future==0.18.3
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
multidict==6.0.4
pycparser==2.21
python-telegram-bot==20.6
requests==2.31.0
sniffio==1.3.0
tornado==6.3.3
urllib3==2.0.7
yarl==1.9.2
